test litellm proxy provider
